Change type of `dynamic_size`

+++ include/dr/constants.hpp
@@ -11,6 +11,6 @@ inline constexpr Index invalid_index{~0};
 template <typename T>
 inline constexpr bool always_false{false};
 
-inline constexpr int dynamic_size = -1;
+inline constexpr isize dynamic_size = -1;
 
 } // namespace dr
